requirements.txt
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
Pillow>=10.4.0
seaborn==0.13.2
streamlit==1.37.1
streamlit_option_menu==0.4.0
